Ensure exactly once initialization of CommonMessageCoordinator (#4273)
* Ensure exactly once initialization of CommonMessageCoordinator

The CommonMessageCoordinator requires an asynchronous initialization
step. Since constructors cannot return promisses, the initialization is
delegted to the asynchronous method `initialize`.
`WebviewIPyWidgetCoordinator` takes advantage of the two-step creation
(`constructor` + `initialize`) to use the
`CommonMessageCoordinator::postEmitter` member between construction and
initialization.

This pattern requires a few safeguards to ensure correct use of
`CommonMessageCoordinator` instances by preventing prevent accidental
errors such as:
- forgetting to call `initialize`
- calling `initialize` more than once
- using the object before initialization

Use the factory method pattern to create initialized instances of
the CommonMessageCoordinator to ensure the initialization always takes
place. Make the `initialize`method private to avoid double
initialization.

Allow the caller to customize the internal `postEmitter` (an
instance of `EventEmitter<{message: string; payload: any}>`) through
dependency injection in the  constructor and factory method. This way,
the `WebviewIPyWidgetCoordinator` can interact with the emitter without
waiting for the initialization to finish.

* Race on the newly created CommonMessageCoordinator