Remove direct usages of the url package
Support for URL is good in all environments, it was actually only being
referenced directly in .node.ts files and it's been available in node
since v10. The package itself also has deprecated url.parse with advice
to use the standard API instead.

It is still imported in tests only via rewiremock.

Part of #11996