fix: first try widget downloads with custom fetch (#17098)
* fix: first try widget downloads with custom fetch

Initially, I set out to fix a logging bug; if the HTTP failed but the custom fetch succeeded, the HTTP error would still be logged. After making that fix, I realized that the logic should maybe be tweaked a bit to instead first try with the custom fetch, then fallback to without.

With minimal context of the existing code, I can't easily determine... but for reviewers with more background, when would using the custom fetch ever fail? Do we even need the fallback? Is the fallback behaviour I'm proposing here desireable or would you still prefer we issue both requests simultaneously and I instead change this PR to only log errors if both fail?

I looked to write or modify a test, but the `RemoteIPyWidgetScriptManager` doesn't appear to have existing unit test coverage. If reviewers feel the change necessitates a test, I can look to cover the component. Though that's likely a bit out of scope.

* fix typo in comment

---------

Co-authored-by: Don Jayamanne <don.jayamanne@outlook.com>